+++ source/module_lr/potentials/xc_kernel.h
